backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
